Fixed spelling error for naming of parent directory and reduced duplicate files for custom data (#101)
