+++ samples/CosmosDBIntegration/ReadThroughSamples/Properties/launchSettings.json
