+++ samples/CosmosDBIntegration/ReadThroughSamples/Properties/serviceDependencies.json
