+++ samples/CosmosDBIntegration/WriteAroundSamples/Properties/launchSettings.json
