+++ samples/CosmosDBIntegration/WriteAroundSamples/Properties/serviceDependencies.local.json
