+++ samples/CosmosDBIntegration/WriteAroundSamples/WriteAroundSamples.csproj
