+++ samples/CosmosDBIntegration/WriteAroundSamples/host.json
