+++ samples/CosmosDBIntegration/WriteBehindSamples/Properties/serviceDependencies.local.json
