+++ samples/CosmosDBIntegration/WriteThroughSamples/Properties/serviceDependencies.json
