+++ samples/CosmosDBIntegration/WriteThroughSamples/Properties/serviceDependencies.local.json
