+++ samples/CosmosDBIntegration/WriteThroughSamples/WriteThroughSamples.csproj
