--- samples/CosmosDbIntegtation/WriteAroundSamples/Models/RedisData.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Microsoft.Azure.WebJobs.Extensions.Redis.Samples
-{
-    public record CosmosDBListData
-    (
-        string id,
-        List<string> value
-    );
-}
